ZavrsniDemo/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
Django==5.1
filelock==3.15.4
pipenv==2024.0.1
platformdirs==4.2.2
pycryptodome==3.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
